service/src/main/java/com/macys/selection/xapi/list/client/response/fcc/ColorwayImageResponse.java
package com.macys.selection.xapi.list.client.response.fcc;

import com.google.common.base.MoreObjects;
import org.apache.commons.lang3.builder.EqualsBuilder;
import org.apache.commons.lang3.builder.HashCodeBuilder;
import org.codehaus.jackson.annotate.JsonIgnoreProperties;

@JsonIgnoreProperties(ignoreUnknown = true)
public class ColorwayImageResponse {

    private Integer seqNumber;
    private String imageName;

    public Integer getSeqNumber() {
        return seqNumber;
    }

    public void setSeqNumber(Integer seqNumber) {
        this.seqNumber = seqNumber;
    }

    public String getImageName() {
        return imageName;
    }

    public void setImageName(String imageName) {
        this.imageName = imageName;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        ColorwayImageResponse that = (ColorwayImageResponse) o;
        return new EqualsBuilder()
                .append(seqNumber, that.seqNumber)
                .append(imageName, that.imageName)
                .isEquals();
    }

    @Override
    public int hashCode() {
        return new HashCodeBuilder(17, 37)
                .append(seqNumber)
                .append(imageName)
                .toHashCode();
    }

    @Override
    public String toString() {
        return MoreObjects.toStringHelper(this)
                .add("seqNumber", seqNumber)
                .add("imageName", imageName)
                .toString();
    }
}
